+++ internal/feed/service.go
@@ -12,7 +12,6 @@ import (
 	coresdk "github.com/goverland-labs/core-web-sdk"
 	"github.com/goverland-labs/core-web-sdk/feed"
 	"github.com/goverland-labs/inbox-api/protobuf/inboxapi"
-	events "github.com/goverland-labs/platform-events/events/inbox"
 	"github.com/rs/zerolog/log"
 	"google.golang.org/grpc"
 	"gorm.io/gorm"
@@ -22,12 +21,7 @@ import (
 
 const (
 	maxPrefillElements = 200
-	minPrefillElements = 5
 )
-
-type Publisher interface {
-	PublishJSON(ctx context.Context, subject string, obj any) error
-}
 
 type SubscriptionsFinder interface {
 	FindSubscribers(ctx context.Context, in *inboxapi.FindSubscribersRequest, opts ...grpc.CallOption) (*inboxapi.UserList, error)
@@ -38,15 +32,13 @@ type Service struct {
 	repo          *Repo
 	subscriptions SubscriptionsFinder
 	sdk           *coresdk.Client
-	publisher     Publisher
 }
 
-func NewService(repo *Repo, subscriptions SubscriptionsFinder, sdk *coresdk.Client, p Publisher) *Service {
+func NewService(repo *Repo, subscriptions SubscriptionsFinder, sdk *coresdk.Client) *Service {
 	return &Service{
 		repo:          repo,
 		subscriptions: subscriptions,
 		sdk:           sdk,
-		publisher:     p,
 	}
 }
 
@@ -63,7 +55,6 @@ func (s *Service) Process(ctx context.Context, item Item) error {
 		return err
 	}
 
-	var pushTitle, pushBody string
 	for _, sub := range resp.Users {
 		subscriberID, err := uuid.Parse(sub.GetUserId())
 		if err != nil {
@@ -100,56 +91,9 @@ func (s *Service) Process(ctx context.Context, item Item) error {
 		if err := s.repo.CreateOrUpdate(&personalized); err != nil {
 			return fmt.Errorf("unable to save feed item '%s' for subscriber '%s': %w", personalized.ID, sub.GetUserId(), err)
 		}
-
-		if !item.AllowSending() {
-			continue
-		}
-
-		if pushTitle == "" {
-			di, err := s.sdk.GetDao(ctx, item.DaoID)
-			if err != nil {
-				log.Error().Err(err).Msgf("core client: get dao: %s: %s", item.DaoID.String(), err.Error())
-				continue
-			}
-			pushTitle = fmt.Sprintf("%s: %s", di.Name, convertAction(item.Action))
-		}
-
-		if pushBody == "" {
-			pr, err := s.sdk.GetProposal(ctx, item.ProposalID)
-			if err != nil {
-				log.Error().Err(err).Msgf("core client: get proposal: %s: %s", item.ProposalID, err.Error())
-				continue
-			}
-			pushBody = pr.Title
-		}
-
-		// todo: send image url here
-		if err = s.publisher.PublishJSON(ctx, events.SubjectPushCreated, events.PushPayload{
-			Title:  pushTitle,
-			Body:   pushBody,
-			UserID: subscriberID,
-		}); err != nil {
-			log.Error().
-				Err(err).
-				Fields(map[string]string{"user_id": subscriberID.String()}).
-				Msg("publish push notification")
-		}
 	}
 
 	return nil
-}
-
-func convertAction(action Action) string {
-	switch action {
-	case ProposalCreated:
-		return "New proposal created"
-	case ProposalVotingQuorumReached:
-		return "Quorum reached"
-	case ProposalVotingEndsSoon:
-		return "Vote finishes soon"
-	}
-
-	return ""
 }
 
 func (s *Service) MarkAsReadByID(ctx context.Context, subscriberID uuid.UUID, id ...uuid.UUID) error {